+++ tests/tests/generic_numeric_slice_xor/generic_numeric_slice_xor.gs.ts
@@ -13,7 +13,7 @@ export function xorLoop<T>(__typeArgs: $.GenericTypeArgs | undefined, dst: $.Sli
 
 export async function main(): globalThis.Promise<void> {
 	let dst: $.Slice<number> = $.arrayToSlice<number>([$.uint(0, 8), $.uint(0, 8)])
-	xorLoop({T: { type: { kind: $.TypeKind.Basic, name: "int" }, zero: () => 0 }}, dst, $.arrayToSlice<number>([$.uint(1, 8), $.uint(3, 8)]), $.arrayToSlice<number>([$.uint(2, 8), $.uint(1, 8)]))
+	xorLoop({T: { type: { kind: $.TypeKind.Basic, name: "uint8" }, zero: () => 0 }}, dst, $.arrayToSlice<number>([$.uint(1, 8), $.uint(3, 8)]), $.arrayToSlice<number>([$.uint(2, 8), $.uint(1, 8)]))
 	$.println($.uint(dst![0], 8), $.uint(dst![1], 8))
 }
