+++ tests/tests/receiver_binding/storage.gs.ts
@@ -72,7 +72,7 @@ export class storage {
 		() => new storage(),
 		[{ name: "IsEmpty", args: [], returns: [] }, { name: "Len", args: [], returns: [] }, { name: "Name", args: [], returns: [] }, { name: "SetName", args: [], returns: [] }, { name: "Truncate", args: [], returns: [] }],
 		storage,
-		{"bytes": { kind: $.TypeKind.Slice, elemType: { kind: $.TypeKind.Basic, name: "int" } }, "name": { kind: $.TypeKind.Basic, name: "string" }}
+		{"bytes": { kind: $.TypeKind.Slice, elemType: { kind: $.TypeKind.Basic, name: "uint8" } }, "name": { kind: $.TypeKind.Basic, name: "string" }}
 	)
 }
