+++ tests/tests/slice_type_assertion/slice_type_assertion.gs.ts
@@ -5,7 +5,7 @@ import * as $ from "@goscript/builtin/index.js"
 
 export async function main(): globalThis.Promise<void> {
 	let v: any = $.interfaceValue<any>($.arrayToSlice<number>([$.uint(7, 8), $.uint(8, 8), $.uint(9, 8)]), "[]byte")
-	let b: $.Slice<number> = $.mustTypeAssert<$.Slice<number>>(v, { kind: $.TypeKind.Slice, elemType: { kind: $.TypeKind.Basic, name: "int" } })
+	let b: $.Slice<number> = $.mustTypeAssert<$.Slice<number>>(v, { kind: $.TypeKind.Slice, elemType: { kind: $.TypeKind.Basic, name: "uint8" } })
 	$.println("byte slice:", $.len(b), $.uint(b![0], 8), $.uint(b![2], 8))
 }
