+++ tests/tests/type_missing_imports/type_missing_imports.gs.ts
@@ -44,7 +44,7 @@ export class file {
 		() => new file(),
 		[],
 		file,
-		{"name": { kind: $.TypeKind.Basic, name: "string" }, "data": { kind: $.TypeKind.Slice, elemType: { kind: $.TypeKind.Basic, name: "int" } }}
+		{"name": { kind: $.TypeKind.Basic, name: "string" }, "data": { kind: $.TypeKind.Slice, elemType: { kind: $.TypeKind.Basic, name: "uint8" } }}
 	)
 }
